// src/app/model/model.ts
export interface Project {
  title: string;
  description: string[];
  image: string;
  url: string;
  tech: string[];
}

export interface Experience {
  title: string;
  icons: string[];
  companies: string[];
  startDate: string;
  endDate: string;
  description: string[];
  skills: string[];
  achievements: string[];
  tech: string[];
}